+++ api/plaid/user.js
@@ -1,5 +1,8 @@
 import { handleDeletePlaidWorkspace } from "../../server/plaid/handlers.js";
+import { applySecurityHeaders, assertMethod } from "../../server/http/responseHelpers.js";
 
 export default function handler(request, response) {
+  applySecurityHeaders(response);
+  if (!assertMethod(request, response, "DELETE")) return;
   return handleDeletePlaidWorkspace(request, response);
 }
